wifi api 페이징 관련 기능 추가 및 변경
- wifi api total type Long에서 int로 변경
- 메서드명 getWifiTotalCountAndQueryCount 변경
- 페이징 처리를 위한 더미 호출 기능 추가

+++ src/main/java/com/zerobase/api/parser/okhttp3/GsonWifiApiMessageParser.java
@@ -42,8 +42,8 @@ public class GsonWifiApiMessageParser implements WifiApiMessageParser {
         return gson.fromJson(element.getAsJsonObject().get("RESULT"), WifiApiResultResponse.class);
     }
 
-    private Long getTotalCount(JsonElement element) {
-        return gson.fromJson(element.getAsJsonObject().get("list_total_count"), Long.class);
+    private Integer getTotalCount(JsonElement element) {
+        return gson.fromJson(element.getAsJsonObject().get("list_total_count"), Integer.class);
     }
 
     private List<WifiApiIntoDetailResponse> getDetailList(JsonElement element) {
